fixes 1 part
