+++ app/src/main/res/layout/fragment_cocktail_details.xml
@@ -8,7 +8,7 @@
     android:background="#380106"
     android:padding="15dp"
     android:layout_margin="5dp"
-    tools:context=".ui.CocktailDetails">
+    tools:context=".ui.CocktailDetailsFragment">
 
     <LinearLayout
         android:layout_width="match_parent"
